refactor: move v2 transaction store method back to common handler
