+++ packages/core-api/lib/versions/2/handlers/transactions.js
@@ -3,9 +3,12 @@
 const Boom = require('boom')
 
 const { TRANSACTION_TYPES } = require('@arkecosystem/crypto').constants
+const { TransactionGuard } = require('@arkecosystem/core-transaction-pool')
+
 const container = require('@arkecosystem/core-container')
 const config = container.resolvePlugin('config')
 const database = container.resolvePlugin('database')
+const logger = container.resolvePlugin('logger')
 const transactionPool = container.resolvePlugin('transactionPool')
 
 const utils = require('../utils')
@@ -24,6 +27,67 @@ exports.index = {
     const transactions = await database.transactions.findAll(utils.paginate(request))
 
     return utils.toPagination(request, transactions, 'transaction')
+  }
+}
+
+/**
+ * @type {Object}
+ */
+exports.store = {
+  /**
+   * @param  {Hapi.Request} request
+   * @param  {Hapi.Toolkit} h
+   * @return {Hapi.Response}
+   */
+  async handler (request, h) {
+    if (!transactionPool) {
+      return {
+        data: []
+      }
+    }
+
+    /**
+     * Here we will make sure we memorize the transactions for future requests
+     * and decide which transactions are valid or invalid in order to prevent
+     * duplication and race conditions caused by concurrent requests.
+     */
+    const { valid, invalid } = transactionPool.memory.memorize(request.payload.transactions)
+
+    const guard = new TransactionGuard(transactionPool)
+    guard.invalid = invalid
+    await guard.validate(valid)
+
+    if (guard.hasAny('accept')) {
+      logger.info(`Received ${guard.accept.length} new transactions`)
+
+      await transactionPool.addTransactions(guard.accept)
+
+      transactionPool.memory
+        .forget(guard.getIds('accept'))
+        .forget(guard.getIds('excess'))
+    }
+
+    if (!request.payload.isBroadCasted && guard.hasAny('broadcast')) {
+      await container
+        .resolvePlugin('p2p')
+        .broadcastTransactions(guard.broadcast)
+    }
+
+    return {
+      data: {
+        accept: guard.getIds('accept'),
+        excess: guard.getIds('excess'),
+        invalid: guard.getIds('invalid')
+      }
+    }
+  },
+  options: {
+    validate: schema.store,
+    plugins: {
+      pagination: {
+        enabled: false
+      }
+    }
   }
 }
